# Added chimeric and unique modes from Perl version