… testdata/fastas/hg19.fa.csv
… testdata/integrationtestcases/Simple;hg19.fa;Colo829-18_S3-sort.bam;chr1;9922-10122;--chimeric -p;.txt
… testdata/integrationtestcases/Simple;hg19.fa;L861Q.bam;chr7;55259400-55259600;-u.txt
… testdata/integrationtestcases/Simple;hg19.fa;L861Q.bam;chr7;55259400-55259600;.txt
… testdata/integrationtestcases/Somatic;hg19.fa;Colo829-18_S3-sort.bam-Colo829-19_S4-sort.bam;chr10;1368753-1369353;.txt
… testdata/integrationtestcases/Somatic;hg19.fa;Colo829-18_S3-sort.bam-Colo829-19_S4-sort.bam;chr10;684892-685492;.txt
… testdata/integrationtestcases/Somatic;hg19.fa;Colo829-18_S3-sort.bam-Colo829-19_S4-sort.bam;chr1;755917-756517;.txt