### Description
There are changes in Perl version from commit https://github.com/AstraZeneca-NGS/VarDict/commit/97e41ebd05cf5469c9d74d2f509024197f598d04, that must be implemented in Java in the first place: 
`-u` (option to process overlapped portion of mate pairs only once) 
`--chimeric` (chimeric artifacts from library constructions, default run). 

### Implementation

- Added options `-u` and `--chimeric` and implemented logic for them. Old logic of overlapping reads was changed in the same way as it was implemented by Zhongwu Lai in commit above.
- Changed method `isGoodVar()` for detection of good variations (also from Zhongwu Lai commit).
- The `goodq` is now the floating-point number (`-q` option, default `22.5`).
- Added new integrations tests for this cases and fixed old tests, now the results are closer to Perl version, but there are still many differences between Perl and Java versions, that need to be implemented later.

### Performance measurements 
All Java measurements are made with -th 8 option. 

File | Master Java before changes | New Java without options | New Java with -u | New Java with --chimeric | New Java with -u and --chimeric | Perl without options
-- | -- | -- | -- | -- | -- | --
AURA_26-ready.bam1,5 Gb |  5 min 10.31 sec | 6 min 34.64 sec | 5 min 6.39 sec | 6 min 33.07 sec | 4 min 55.72 sec | 11 hours 25 minutes

The performance for -u mode is now the same as was before (when this mode was by default).