@mjafin 2018-05-14
Just to understand this correctly, with `-u` only one read will be processed from an overlapping pair (as here https://github.com/AstraZeneca-NGS/VarDictJava/issues/60) and without `-u` we revert to old way of ignoring the fact that the reads are overlapping (and therefore count twice)?