@PolinaBevad 2018-05-14
@mjafin , Exactly! And also the overlap detection algorithm is now identical to perl detection.