@mjafin 2018-05-14
Could you summarise how the logic is different to the previous implementation please?