@PolinaBevad 2018-05-14
@mjafin , in the previous implementation read was skipped if it was in forward direction, the mate of read was in reverse direction, start of mate was less than or equal to read end and the start of read was more than or equal to the start of mate. Also chromosomes of mate and read must be the same. 

Now read is skipped only if unique mode enabled and read is in forward direction with start of read more than or equal to the start of mate. Also it is skipped now at the very beginning of the processing Cigar.

Perhaps there is a question about the same chromosomes, but it is not checked in perl now. Do you think we must leave this check?