@mjafin 2018-05-14
I'm struggling to understand how the changes affect the logic. Maybe Zhongwu @zhongwulai could comment on why the Perl fix was different from Java fix?

Further, does it address the strand bias issue discussed here (I suspect not): https://github.com/AstraZeneca-NGS/VarDictJava/issues/102