@PolinaBevad 2018-05-14
@mjafin , of course, let's wait for the answer of Zhongwu.  
The strand bias issue for variation in the https://github.com/AstraZeneca-NGS/VarDictJava/issues/102 is fixed for this concrete variation (T->A), but because of another changes in perl version: now frequency with bias (2;1) must be more than 20 (was 25) to detect it as good var.