@zhongwulai 2018-05-18
@PolinaBevad You're right, the chromosome check should be added ($a[6] eq "=").   So please go ahead and add the check.  When -u option is enabled, it doesn't skip the read completely, it only skips the overlapping portion in the first read.  The part I don't understand is why the performance decreases significantly.