@PolinaBevad 2018-05-21
@zhongwulai , thanks! I added this check.
We also found that there are some problems with java optimizations on large methods (in this case method `parseSam()`), so some refactoring was done to avoid this and now the performance is almost the same as before. The new measurements are given in the table.