@PolinaBevad 2018-05-21
@mjafin, can we merge this PR or do you have any other comments?