@zhongwulai 2018-05-22
I'm OK with current implementation.  Many thanks!